Update SAW correctness proof of AES to use T-boxes
This proof uses a Cryptol specification based on a T-box encoding
instead of an S-box encoding, matching the underlying C code. It
currently covers only AES-128, but AES-256 will come soon. It takes
about 0.5 hours to prove correctness of encryption on a 3-year-old
MacBook Pro, and about 2.5 hours for decryption.